+++ common/src/main/java/com/pickcar/security/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.pickcar.security.config;
 
-import com.pickcar.jwt.JwtAuthenticationFilter;
+import com.pickcar.security.filter.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,7 +20,11 @@ public class SecurityConfig {
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        //TODO: authorizeHttpRequests에 제어할 경로 설정
+        /*
+            1. TODO: 로그인 form 비활성화
+            2. 제어할 경로 설정(권한 설정)
+            3. 기타 등등... 하면서 추가하기
+        */
         return http
                 .csrf(csrf -> csrf.disable())
                 .authorizeHttpRequests(auth -> auth
