+++ common/src/main/java/com/pickcar/security/filter/JwtAuthenticationFilter.java
@@ -1,5 +1,7 @@
-package com.pickcar.jwt;
+package com.pickcar.security.filter;
 
+import com.pickcar.security.jwt.JwtProvider;
+import com.pickcar.security.jwt.UserPrincipal;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
