+++ common/src/main/java/com/pickcar/security/jwt/UserPrincipal.java
@@ -1,4 +1,4 @@
-package com.pickcar.jwt;
+package com.pickcar.security.jwt;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
